fix(move-vault): delete vault conditionally

+++ feature/vault/src/main/kotlin/de/davis/keygo/feature/vault/presentation/VaultFlowViewModel.kt
@@ -208,7 +208,8 @@ internal class VaultFlowViewModel(
             moveItemsToVault(switcher.srcVaultId, dstVaultId) { progress ->
                 moveProgress.update { progress }
             }.onSuccess {
-                viewModelScope.launch { deleteVault(switcher.srcVaultId) }
+                if (moveState.delete)
+                    viewModelScope.launch { deleteVault(switcher.srcVaultId) }
             }
             dismiss(force = true)
         }
